requirements.txt
torch
torchvision
scikit-image
numpy
pillow
flask
boto3
sagemaker
pytest
pytest-cov